Ibft validates block on proposal reception (#583)
Previously, IBFT only validated the header of the block on reception
of a proposal. However in order to minimise the risk of a "bad block"
being detected after a PreparedCertificate being created, it has been
decided to validate the whole block (i.e. exercise the transactions,
process the block and ensure hashes etc line up with the header).

There is still some (minor) risk of a failed block import, however
it is substantially reduced.