Use absolute data-path instead of relative. (#894)
When using the data-path internally use an absolute version instead of the path instead of a relative form.  

This fixes a null pointer exception when the data-path is set to an empty string (in a TOML config using `data-path=""` or CLI using `--data-path=`) and interprets that empty string to be the current working directory.
Signed-off-by: Adrian Sutton <adrian.sutton@consensys.net>