RLP encode subcommand (#944)
Add RLP encoding subcommand for at least ibftExtraData to be inserted in genesis file.

There's only one encodable type for the moment, however the `--type` option is already there with a default value on this type. We'll add more possible types later if needed and the current type will remain the default one.

The file output is optional and it writes to standard output by default, same for
standard input or a file.

Tests and doc change.
Signed-off-by: Adrian Sutton <adrian.sutton@consensys.net>