Add Unstable Options to the CLI (#1213)
Adds handling for options that need to be configurable for tuning and performance but in general are not long-term supported configuration options.

Options that are added via a mixin, are hidden, and that start with `--X` are considered 'Unstable Options.'  These options do not show up via normal help facilities but are enumerated via a `XHelp` CLI invocation.  These options are unstable, unsupported, and undocumented.  They may be added, removed, or modified without prior warning or notice.  Use them at your own risk.

For example, the `--Xrocksdb-max-open-files` flag can be set via CLI and can increase the number of max open files RocsDB uses.  On a 8GB memory system the default is appropriate, but on a 128GB memory system a much larger number may improve performance.  Normal operation scenarios will not need to modify this flag.
Signed-off-by: Adrian Sutton <adrian.sutton@consensys.net>