[PAN-2703] JSON-RPC api net_services should display the actual ports (#1628)
Currently when json-rpc and ws services are configured with a port value of 0, the "net_services" endpoint will display their ports as "0".  A configured port of "0" actually means "choose a random port".  We need to get the actual port that those services are using displayed in the "net_services" JSON-RPC response.