[PAN-2829] Account versioning (#1612)
Initial EIP-1702 versioning code.  Version internally is an `int` for Java 
reasons. In the future if the versions become sparse or large we will create a 
"proxy" version index by putting the standardized indexes into an enum and 
using the enum ordinal instead.

Reference tests all pass (hence the WorldStateMock if block) and
GenesisStateTest has an explicit version test now.