move prepare to fix race condition
When importing blocks quickly, there's a chance that some state from the
block after the mark block makes it into the database before the node
added listener is fully attached. This means that state after the mark
block could be removed. This creates the node added listener once and
for all before the block added event observer so the race never happens.

Signed-off-by: Ratan Rai Sur <ratan.r.sur@gmail.com>