Repair EthGetLogs returning incorrect results (#128)
An error was detected (PAN-3248) whereby if "Null" appeared in a Log
Topic filter, it and all subsequent filters were lost (and thus
were not used to filter responses) - thus Besu would return too many
results (as the filters were less restrictive than requested).

This was determined to be an issue in the TopicParameterDeserialiser
which is resolved in this commit.

Signed-off-by: Trent Mohay <37158202+rain-on@users.noreply.github.com>
Signed-off-by: edwardmack <ed@edwardmack.com>