Remove LogSeries Object (#289)
Remove the LogSeries object and replace it with a plain old List<Log>.
The subclass had two features beyond a simple list. First it would
aggregate the logs as they were added into a bloom filter, but that
bloom filter was never used.

Second it had a RLP encoding convenience method. In the few cases this
is used the RLP encoding is unwrapped.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>