Add Tuweni to Plugin-APIs
Generally, byte[] -> Bytes of some form.  Most of the changes are the
side effect of the type changes or chaning to the names of Tuweni
equivilant calls (getHexString->toHexString, etc).

UnformattedData -> Bytes
Log Topics went from Hash to Bytes32
Difficulty went to UInt256 to match core impl.
Quantity lost BinaryData and is just getValue() and toHexString()

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>