Tracing - respect tx boundary for EIP-2200 cost calculations
When we calculate the cost of a SSTORE in trace we need to consider the
original value as of when the last transaction finished, not when the
whole block started.

To create a test for this the test blockchain also needed to be upgraded
to Istanbul.  This resulted in a large number of cost only changes.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>