Reduce Trace Memory Consumption
Swith the EVM Memory to a copy on write structure, so when the trace
reads the memory they don't have to copy mutable bytes but can instead
simply use the value as immutable.

Also, fix a state diff bug when an account is deleted that was also just
created in the TX don't report it as an all equal diff.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>