Update Code of Conduct and Security
This Bring our code of conduct and security file in line with automated
checks.

The changes to `SECURITY.md` are strictly formatting, adding newlines so
it line wraps at 80 lines.

The `CODE_OD_CONDUCT.md` is a more substantial change.  Rather than the
being subject to both the contributor covenant and the Hyperledger CoC
this change makes the project only subject to the Hyperledger CoC.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>