Loggers should be private, static and final.  Enforced with errorprone
Some of our loggers were not private, static, and final.  In a few
cases these were non-static fields in classes that were repeatedly
instantiated in core transaction logic.

This is enforced via a new ErrorProne check, so the PR includes fixes
for all of the places this was a problem, not just the performance
impacting code.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>