Increase Netty threadpool size for handshake handling
It was found that all ECIES handhshaking is taking place in
an NioEventLoop of being seviced by one worker.

This in turn meant that, during periods of high demand,
many potential peers were timing out (as the latter peers
had to wait for earlier peers to complete prior to
being serviced).

This change grows the threadpool to 10 (arbitrarily selected),
which minimises the number of peers waiting for connectivity.

Signed-off-by: Trent Mohay <trent.mohay@consensys.net>