Caching Block Bodies in BlockchainQueries
In situations where we are doing a lot of queries in the same block
(like all receipts in a block, one query at a time) we wind up
re-loading and re-parsing the same block for each query.  Use a local
cache of soft values.  Being soft values they will clear if there is
memory pressure.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>