Remove use of Generics in ProtocolContext
ProtocolContext uses a geneic for the consensus state, which has a very
large footprint across the code to accomplish what it intends to
accomplish.

Instead replace it with java.lang.Class#cast, which provides runtime
security, and use generics to provide the compile time sugar that
allows for chained methods of the appropriate type.  Then remove it's
(quote large) footprint from the rest of the code.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>