Remove use of generics in ProtocolContext consensus object (#1039)
ProtocolContext uses a generic for the consensus state, which has a very
large footprint across the code to accomplish what it intends to
accomplish.  For every call there are about 61 other lines per call that 
need to be updated, over 1300 lines total.

Instead replace it with java.lang.Class#cast, which provides runtime
security, and use generics to provide the compile time sugar that
allows for chained methods of the appropriate type.  Then remove its
(quite large) footprint from the rest of the code.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>