Add 0x check for Address, Bytes, and Bytes32 in GraphQl requests
Produce error response if any of Address, Bytes or Bytes values
inside GraphQl requests are not prefixed with 0x as defined by the spec.

Fix old tests that used values that are not prefixed with 0x and add
additional tests to confirm Address, Bytes and Bytes32 values will be
rejected if not prefixed with 0x.

Signed-off-by: Mak Muftic <mak@nodefactory.io>