Fix sync timeout issue  (#1149)
This PR fixes an error when downloading chain (Async operation failed)

When there are several blocks which are very large (> 12M) and which are requested in the same segment (by default 200) we can have timeouts and never manage to synchronize. This modification will make it possible to gradually reduce the size of the segment with each attempt. Then the segment resumes its default size for the next blocks

If the reduction is not enough at the last attempt we try with a single block

Signed-off-by: Karim TAAM <karim.t2am@gmail.com>
(cherry picked from commit 17556485f366ca4743747058b411f7581b094e04)
Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>