Copy Subscribers to Prevent Unwanted Recursion (#1169)
Occasionally, a callback given to a subscriber will complete a future 
which then runs another task which adds another callback to the 
subscriber list of the same type. By the time the first callback is 
finished, the new callback (which presumably wants another "real" 
instance of the event to happen) could be next up in the list of the 
first's subscribers callback iteration. This takes a snapshot of the 
subscribers when the desired event happens so there's no risk of 
following down the rabbit hole.

I collected some rough data about the size of the subscriber lists to 
see the penalty for copying them and I found that by the time a full 
sync began or a fast sync had started downloading headers, the maximum 
subscribers any one Subscribers class had was 7.

Signed-off-by: Ratan Rai Sur <ratan.r.sur@gmail.com>