Fix GraphQL IBFT Coinbase Query
When IBFT produces blocks the coinbase may correspond to an empty
account (as mining rewards are not paid out).  In this case allow
GraphQL to return an all zero account for the empty account only in
context of a miner.  Other tests exist that verify a plain accounts
query of a non-existent account throws an exception.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>