Experimental Backup and Restore (#1235)
Experimental backup and restore via `operator x-backup-state` and `operator x-restore-state` CLI commands.  Besu needs to be in a non-operational state for the backup and restore to occur.  Restore has only been tested into an empty database, not on top of an existing database.

This feature is not advised for production use at the moment.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>