Attempt Pending Request to BUSY && ASSIGNED && DISCONNECTED Peer (#1321)
Any process that depends on pre-assigned peer tasks is liable to grow 
our pendingRequests out of control.

When we attempt to execute a pending transaction, we check for that 
peers availability:

PendingPeerRequest.java
```
     final Optional<EthPeer> selectedPeer = 
         
leastBusySuitablePeer.filter(EthPeer::hasAvailableRequestCapacity); 
  
     selectedPeer.ifPresent(this::sendRequest); 
     return selectedPeer.isPresent(); 
   } 
 }
``` 
However, if that peer was disconnected while there were max outstanding 
requsts, we'll never reattempt it as selectedPeer will be empty. 
Furthermore, we'll return false and keep that pending request in our 
list of requests.

With this change we indicate that we aren't waiting for any outstanding 
requests when we disconnect from a peer, allowing the existing logic 
that triggers when peers are disconnected to be hit.

The getter I added just for testing purposes was to get around the 
inadequacies of the MockPeerConnection. There's a whole lot of callback 
plumbing that's needed for a simple ethPeer.disconnect to do the same 
thing it does when we call it in our production code.

Signed-off-by: Ratan Rai Sur <ratan.r.sur@gmail.com>