better disconnection handling (#1348)
This solves the same problem as 3a39fbb but better. The attempt 
there has an issue where it still emulates sending the request and 
therefore only 5 requests max can be aborted. Thus the problem of the 
peer becoming stuck in a busy state remains. Here we do what we 
should've done all along, which is proactively abort these requests 
when a disconnect event occurs.

Signed-off-by: Ratan Rai Sur <ratan.r.sur@gmail.com>