Add gas measurement to full import log line
To provide more context to admins and to add more meaning the full sync
log lines report the amount of gas processed in MegaGas/Second.  Some
blocks are fuller than others and this provides a gauge as to wether the
import is slowing down or if the blocks are filling up.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>