Remove Unnecessary LimitedNewPooledTransactionHashesMessage (#1511)
The reason this existed in the first place was only because the author 
wanted to impact the existing code as little as possible and so copied 
analogous classes. However, this class didn't make sense in the context 
of the eth65 changes because there isn't size-in-bytes limiting logic, 
only an implementaion-specific max-count logic that can easily fit in 
the sender class. The tests were deleted because we already have 
coverage of the 4096 batch size in PendingTransactionsMessageSenderTest.

Signed-off-by: Ratan Rai Sur <ratan.r.sur@gmail.com>