Ensure peers are validated before sending requests
Don't send requests to peers that have not been fully validated (such as
validating the absence/presence of the DAO block). This has the intended
side effect that unvalidated peers don't report chain height to
prometheus monitoring. On occasion classic/mainnet peers will be
mid-validation when the sync status is requested and the wrong chain
will be reported as the height, typically classic in a mainnet node
because classic is higher.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>