Update eth_call handling of account balance
When calling `eth_call` by default account balances will be ignored when
executing the call. If the user wants the gas balance to be a
consideration in the call a new `strict` param in the call params can be
set to true, which will enforce the balance rules.  This is the same
behavior as is observed in `eth_estimateGas`.

Addresses #502

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>