Bonsai Tries: Respect account deletion across transacitons (#1952)
When in bonsai mode, if inside one block an account is deleted via self
destruct and then the account is re-created the latter sstore operation
reports the last SSTORE value prior to self destruct. Thus if the new
write is the same as the value before delete a non-updating gas cost
will mistakenly be applied.  Thus we first check accountWasDeleted when
reading original slot values.

Found in mainnet block 11905274 for account
0x7f150bd6f54c40a34d7c3d5e9f56.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>
Signed-off-by: Richard Hart <richardhart92@gmail.com>