Less Spotless Gradle Warnings (#2614)
* Less Spotless Gradle Warnings

Spotless and Gradle 7.x does not play nicely with using a fileTree as a
target and should be using plain strings. This should get rid of the
`Execution optimizations have been disabled`... warnings.

* Move Generated Reference Tests into the `build` dir

Generated reference test are currently generated into `src`, which is in
violation of the standard maven layouts.  They should be in the `build`
directory as they are artifacts not maintained in source control.

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>