fix: private contracts not able to call public contracts that call other public contracts (#2816)
Private accounts are not able to change the state of public accounts. When transferValue is called in MessageCallProcessor it attempts to get a mutable account. This is only required when a transfer of value is happening. If a transfer of value from a private contract to a public contract is attempted an error will be thrown.

Signed-off-by: Antony Denyer <git@antonydenyer.co.uk>