Merge pull request from GHSA-7pg2-p5vj-xp5h
Shift operation expect the shift amount to be an unsigned int 256, even
though only 8 bits provides meaningful value. Besu optimized this by
casting this to an int, but it was signed, resulting in a bad shift
amount. To ensure that such errors don't cause issues again test the
binary data types as a potentially negative along the major int
types (8,16,32,64,128,256).

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>