Hacky POC to fix TLS P2P failing to process messages over a certain size (> TLS record limit of 16 KB)
Copied Framer impl into PlainFramer and commented out the encryption-related bits.
This means we still get some of the partial message processing logic.
Essentially, we've added a HEADER containing the expected size of the message and will wait for more data if the message buffer only contains a partial message.

Signed-off-by: Simon Dudley <simon.dudley@consensys.net>