Remove shutdown latch from BftProcessor
We don't want BftMiningCoordinator.stop() to wait for BftProcessor to shutdown before calling shutdown on the BftExecutor.
This prevents BftMiningCoordinator from being able to be shutdown from the same thread.
We want to enable shutdown from the same thread to support migrating from one miningCoordinator instance to another (IBFT -> QBFT).

Signed-off-by: Simon Dudley <simon.dudley@consensys.net>