Remove Gas object and replace with primitive long
Remove the Gas object from the EVM and replace it with the primitive
long. This will have positive impact on short lived object garbage
collection stats, which at very high load causes significant performance
issues.

This also codifies EIP-4803 in the Besu EVM, limiting gas to a signed
64-bit long internally.

There was one notable issue with this transition. Some reference tests
push the limits of memory expansion which results in gas costs over the
EIP limit. The solution is to "clamp" such addition and multiplication
operations at the max or min long value, resulting in the out-of-gas
check failing (such operations will never have a valid execution with
max gas remaining).

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>