Split FastSyncState into PivotBlockProposal and PivotHolder
Turns out the FastSyncState was used for two purpouses that were working against each other. Firstly it was used as a holder of current pivot with an option to update the pivot.
Secondly it was used to keep a proposal for a new pivot - as in a number or hash without the knowledge of header. When we would have such a proposal we would then imediatelly go and ask for an actual header for the pivot.

This commit separates this two functionalities into two. The convolution of those two touched lots of places that now may be simplified slightly.
As a consequence the SnapSyncState does not extend from FastSyncState anymore (FastSyncState does not exist anymore).

Signed-off-by: Jiri Peinlich <jiri.peinlich@gmail.com>