Add checkpoint sync (#3849)
Add a new way to synchronize which is X_CHECKPOINT. This mode is experimental so use it at your own risk. This mode allows you to do like a snapsync but starting from a specific checkpoint instead of starting from the genesis.

This checkpoint will be in the genesis configuration of each network. To add the checkpoint mechanism in a network you just have to add the checkpoint section in the genesis.

Currently there is a checkpoint for ropten, goerli and mainnet.

Mainnet on i3.2xlarge <6 hours
Goerli on i3.2xlarge <1 hours

Signed-off-by: Karim TAAM <karim.t2am@gmail.com>