Accept empty header set in range headers validation (#4189)
If a response to the get header P2P request only returns the header that
is the start of the range we may need to trim it to an empty response,
this is breaking the validation response. One client has started
returning only the range header start in some circumstances (which is a
valid response per spec). Because we sometimes request an overlapping
header this results in an empty stream once we cut the duplicates.

fixes #3336

Signed-off-by: Danno Ferrin <danno.ferrin@gmail.com>