Migrate storage APIs to Bytes32 (#4562)
The current internal storage APIs implement UInt256 rather than Bytes32,
which is an accident of prior EVM design.  This migrates all internal
handling of storage addresses and values to Bytes32.

The main performance gain is reduced GC pressure, in that UInt256
stores the data as an array of ints, whereas Bytes32 keeps a sliced or
wrapped array of bytes. Since database APIs are all byte array based
this should remove unneeded conversions when going to/from the DB.

Signed-off-by: Danno Ferrin <danno.ferrin@swirldslabs.com>
Signed-off-by: Sally MacFarlane <macfarla.github@gmail.com>