Optimize edge case when CL sends new payload out of order
It could happens that CL sends new payload out of order, creating
a temporary gap in the chain, this could cause to redownload some block
headers that are already present, and create glitches in the progession log.
This changes, just check if the headers are already present before trying
to download them.

Signed-off-by: Fabio Di Fabio <fabio.difabio@consensys.net>