EIP-4750 EOF Functions (#4781)
Implement CALLF, RETF, and JUMPF operations. This introduces the
notion of "code sections" into the code. For legacy code section zero
consists of all of the code and is the only code section. Also
implements EIP specified code validation to ensure validated EOF code
won't use CALLF or JUMPF to a non-existent section.

Signed-off-by: Danno Ferrin <danno.ferrin@swirldslabs.com>