Add trie log pruning after TrieLogManager.persist
Feature toggled by --Xtrie-log-pruning-enabled
Each time a trie log is persisted, the current trie log is cached and the pruner is run against the oldest entries in the cache.
This makes no attempt to manage the backlog of old trie logs, it will only prune what has been added to the cache, i.e. trie logs that have been added since the feature was enabled.
Pruner limit exists in case of exceptional circumstances, but we should only ever be pruning all the forks for a single block number during each prune execution.

Signed-off-by: Simon Dudley <simon.dudley@consensys.net>