load all accounts into cache (#6101)
There was a slight problem on the bonsai side because all account reads did not go through a single method. One of the two add the account to the cache but the other did not. This had two consequences:

Less good performance because certain accounts had to be read several times and also all account reads were not marked in the trielog. This will fix both problems.

Signed-off-by: Karim TAAM <karim.t2am@gmail.com>