Add PragueEOF fork
Add the "PragueEOF" fork which is prague+EOF. For genesis and evmtool
trace, the Prague fork is just prague. To use EOF for networks and CLI
traces use the 'PragueEOF' fork. For reverence tests the "Prague" fork
maps to "PrageEOF" because reference tests currently have EOF tests
wired into Prague.

Signed-off-by: Danno Ferrin <danno@numisight.com>