Add majority vote rule to removal of maintainers
See the Besu Charter conversation at https://github.com/hyperledger/besu/pull/8716/files#r2121409576

On the most recent community call we agreed to keep maintainer addition/removal rules in one place, so rather than have them in both the charter and the `Maintainers.md`, we we have them in the latter and the charter would refer to it.

This PR updates `Maintainers.md` to include the text from the charter about requiring a majority of maintainers to vote on removal of an existing maintainer.

There is perhaps a follow on question about whether we do indeed want to require a majority approval to remove a maintainer. That would be a higher bar than is required to become a maintainer in the first place (3 votes + 0 vetoes). Should removing a maintainer be the same?

Signed-off-by: Matt Whitehead <matthew.whitehead@kaleido.io>