Move Keccak256 benchmark to JMH and fix JMH integration (#8920)
This PR fixes the JMH integration that Besu already had so that we can write microbenchmarks again. I also added a benchmark for the keccak256 hashing so that I can demo the benchmarking since other benchmarks in there are currently broken and removed the old one from Benchmarks.java.

Profiler only records the measurements and ignores the warmup phase, however it uses the warmup phase to warmup the profiler itself which is great.
One can also add other options to the `async=` command using the `;` as a separator. For a list of available options (might not overlap with the async profiler itself as these need to be mapped jmh -> asyncProfiler)