Downgrade ethereum/tests to v14.1 (#9128)
Revert "Update ethereum/tests submodule to v17.2 for Prague support (#8648)"
This reverts commit f27b7a10b941a5f6cb761228b4b86c9b8643aa6f.

v17.2 moved a significant chunk of tests to LegacyTests which we are not referencing in Besu and so we lose test coverage.

This revert is a quick fix to get more reference test coverage.
Longer term fix is either waiting for EESTs to catch up with bridging the coverage gap.
Or updating our referenceTests config to point to the LegacyTests locations.

Signed-off-by: Simon Dudley <simon.dudley@consensys.net>